refactor(ffmpeg): split hardware detection and ffprobe into modules
P3 后续 — 拆分 FFmpegTool 三职责:

- 新增 video_tools/hardware.py: HWAccelType + detect_hw_accel / 各 _check_*
  / ffmpeg_supports_encoder / get_hw_accel_encoder（FFmpeg 编码器探测经统一
  执行器，nvidia-smi/wmic 保留裸 subprocess）
- 新增 video_tools/probe.py: run_ffprobe_json + get_duration/resolution/
  framerate/bitrate/video_info
- ffmpeg_tool.py 瘦身为业务便捷门面（trim/concat/extract_audio 等保留），
  硬件/探测方法改薄委托；HWAccelType 从 hardware 再导出
- __init__.py 导出 hardware / probe 子模块
- 向后兼容：FFmpegTool.* 调用面与 HWAccelType 导出不变（测试 patch 底座
  SecureExecutor.run，全部照过）；新增 2 个测试覆盖拆出模块

回归: ruff 通过, pytest 597 passed / 11 skipped, mypy 改动文件 0 新增错误
