+++ tests/test_ffmpeg_tool.py
@@ -304,3 +304,35 @@ class TestFFmpegToolIntegration:
 
         for method in required_methods:
             assert hasattr(FFmpegTool, method), f"FFmpegTool.{method} 不存在"
+
+
+class TestSplitModules:
+    """硬件检测 / ffprobe 已拆为独立模块，验证可直接调用 + 委托一致。"""
+
+    def test_hardware_module_functions(self):
+        from scenefab.services.video_tools import hardware
+
+        for fn in (
+            "detect_hw_accel",
+            "get_hw_accel_encoder",
+            "ffmpeg_supports_encoder",
+            "check_nvidia_smi",
+            "check_vaapi",
+            "check_intel_cpu",
+        ):
+            assert callable(getattr(hardware, fn)), f"hardware.{fn} 缺失"
+        # HWAccelType 权威定义在 hardware，FFmpegTool 再导出同一对象
+        from scenefab.services.video_tools.ffmpeg_tool import HWAccelType
+
+        assert HWAccelType is hardware.HWAccelType
+
+    @patch("scenefab.utils.security.SecureExecutor.run")
+    def test_probe_module_delegation(self, mock_run):
+        from scenefab.services.video_tools import probe
+
+        mock_run.return_value = Mock(
+            returncode=0, stdout='{"format": {"duration": "12.5"}}'
+        )
+        # 模块级函数与 FFmpegTool 委托结果一致
+        assert probe.get_duration("/v.mp4") == 12.5
+        assert FFmpegTool.get_duration("/v.mp4") == 12.5
